Fix issue with module-level pass: handle `declare`
Several (all?) testdata files were failing when the
`FunctionTerminationPass` was run via proxy (from the module-level
pass), instead of being run at the function level. Unfortunately, "run
from the module" is the only way we think we can handle the call-graph
level analysis.

The symptom was LLVM stalling out -- looping forever? -- when we asked
for `ScalarEvolutionAnalysis`. I opened up LLDB and stepped through,
and... it was fine! Looking at `simple.c`, the pass produced results for
`get_value` and `main` just fine.

...but our functions are not the only functions listed in a module! For
instance, `simple.c` invokes `malloc`; as a result, `malloc` gets a
`declare`d in our module, without a function body. And *that's* where
the analysis was stalling out -- our pass was running on `F.getName() ==
"malloc"`!

Apparently, function-level passes are only run across functions
`define`d in the module -- but the listing of module contents includes
functions `define`d or `declare`d. I haven't really checked this, but it
is consistent with what we've seen, and would explain the difference
between running in function/module modes.

My fix (this commit) is to return `Unknown` if the function does not
have a body, i.e. has no basic blocks. This gets us back to a passing --
though incorrect -- state.